Pull `BaseCoordinateFrame` into its own module
